requirements.txt
sounddevice
soundfile
pynput
google-generativeai
numpy
vosk